client: Handle multiple Extend API errors
When a particularly bad thing happens, Switchvox's Extend API will return
a list of Errors. Since we convert those Errors into Exceptions, there's
no point in raising an Exception for each. It's also a pain for consumers
to expect both a single instance of an error in the exception, as well as
a list of them (this is already a nuisance on many of the responses). As
such, we just raise an Exception for the first error returned.